update feature folders
