+++ Halcyon.Web/src/components/main-nav.tsx
@@ -19,7 +19,7 @@ const routes: MainNavRoute[] = [
     { href: '/', label: 'Home' },
     { href: '/about', label: 'About' },
     {
-        href: '/user',
+        href: '/users',
         label: 'Users',
         roles: isUserAdministrator,
     },
