+++ Halcyon.Web/src/features/users/hooks/use-get-user.ts
@@ -20,7 +20,7 @@ export const useGetUser = (id: string) => {
     return useQuery({
         queryKey: ['user', id],
         queryFn: ({ signal }) =>
-            apiClient.get<GetUserResponse>(`/user/${id}`, signal, undefined, {
+            apiClient.get<GetUserResponse>(`/users/${id}`, signal, undefined, {
                 Authorization: `Bearer ${accessToken}`,
             }),
     });
