+++ Halcyon.Web/src/features/users/hooks/use-unlock-user.ts
@@ -1,29 +1,21 @@
 import { useMutation, useQueryClient } from '@tanstack/react-query';
 import { useAuth } from '@/components/auth-provider';
 import { apiClient } from '@/lib/api-client';
-import type { Role } from '@/lib/session';
 
-type UpdateUserRequest = {
-    emailAddress: string;
-    firstName: string;
-    lastName: string;
-    dateOfBirth: string;
-    roles?: Role[];
-    version?: number;
-};
+type UnlockUserRequest = { version?: number };
 
-type UpdateUserResponse = {
+type UnlockUserResponse = {
     id: string;
 };
 
-export const useUpdateUser = (id: string) => {
+export const useUnlockUser = (id: string) => {
     const { accessToken } = useAuth();
 
     const queryClient = useQueryClient();
 
     return useMutation({
-        mutationFn: (request: UpdateUserRequest) =>
-            apiClient.put<UpdateUserResponse>(`/user/${id}`, request, {
+        mutationFn: (request: UnlockUserRequest) =>
+            apiClient.put<UnlockUserResponse>(`/users/${id}/unlock`, request, {
                 Authorization: `Bearer ${accessToken}`,
             }),
         onSuccess: (data) => {
