+++ Halcyon.Web/src/features/users/hooks/use-update-user.ts
@@ -1,21 +1,29 @@
 import { useMutation, useQueryClient } from '@tanstack/react-query';
 import { useAuth } from '@/components/auth-provider';
 import { apiClient } from '@/lib/api-client';
+import type { Role } from '@/lib/session';
 
-type UnlockUserRequest = { version?: number };
+type UpdateUserRequest = {
+    emailAddress: string;
+    firstName: string;
+    lastName: string;
+    dateOfBirth: string;
+    roles?: Role[];
+    version?: number;
+};
 
-type UnlockUserResponse = {
+type UpdateUserResponse = {
     id: string;
 };
 
-export const useUnlockUser = (id: string) => {
+export const useUpdateUser = (id: string) => {
     const { accessToken } = useAuth();
 
     const queryClient = useQueryClient();
 
     return useMutation({
-        mutationFn: (request: UnlockUserRequest) =>
-            apiClient.put<UnlockUserResponse>(`/user/${id}/unlock`, request, {
+        mutationFn: (request: UpdateUserRequest) =>
+            apiClient.put<UpdateUserResponse>(`/users/${id}`, request, {
                 Authorization: `Bearer ${accessToken}`,
             }),
         onSuccess: (data) => {
