+++ Halcyon.Web/src/features/users/update-user/update-user-form.tsx
@@ -5,7 +5,7 @@ import { LoadingButton } from '@/components/loading-button';
 import { DateField } from '@/components/form/date-field';
 import { TextField } from '@/components/form/text-field';
 import { SwitchField } from '@/components/form/switch-field';
-import type { GetUserResponse } from '@/features/user/hooks/use-get-user';
+import type { GetUserResponse } from '@/features/users/hooks/use-get-user';
 import { isInPast } from '@/lib/dates';
 import { roleOptions, roles } from '@/lib/session';
